Pesquisa/Gabriel/dataset_textura/requirements.txt
absl-py==1.4.0
accelerate==0.24.1
aiofiles==23.2.1
anyio==3.6.2
apturl==0.5.2
argcomplete==3.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asn1==2.4.2
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
azure-ai-formrecognizer==3.2.1
azure-common==1.1.28
azure-core==1.27.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.10.0
bleach==6.0.0
blinker==1.6.2
boto3==1.34.108
botocore==1.34.108
Brlapi==0.8.3
Brotli==1.1.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
chumpy @ git+https://github.com/mattloper/chumpy@51d5afd92a8ded3637553be8cef41f328a1c863a
click==8.1.3
colorama==0.4.6
comm==0.1.2
command-not-found==0.3
contourpy==1.0.6
cryptography==42.0.7
cupshelpers==1.0
cycler==0.11.0
dacite==1.7.0
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
enum-compat==0.0.3
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==1.2.0
fastjsonschema==2.16.2
fiftyone==0.23.8
fiftyone-brain==0.16.1
fiftyone_db==1.1.2
filelock==3.4.2
Flask==2.0.3
flatbuffers==23.1.21
fonttools==4.38.0
fqdn==1.5.1
fsspec==2023.9.1
ftfy==6.2.0
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.4.0
gbinder-python==1.1.2
glob2==0.7
google-api-core==2.12.0
google-api-python-client==1.12.11
google-auth==2.16.1
google-auth-httplib2==0.1.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.61.0
graphql-core==3.2.3
greenlet==1.1.2
grpcio==1.51.3
h11==0.14.0
h2==4.1.0
h5py==3.8.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
huggingface-hub==0.23.0
humanize==3.14.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.4
imageio==2.34.1
importlib-metadata==4.6.4
inflate64==1.0.0
iopath==0.1.10
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonlines==4.0.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.1
jupyter_core==5.1.5
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kaleido==0.2.1
keras==2.11.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==15.0.6.1
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Markdown==3.4.1
MarkupSafe==2.0.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mongoengine==0.24.2
more-itertools==8.10.0
motor==3.4.0
mpmath==1.3.0
msgpack==1.0.3
msrest==0.7.1
multivolumefile==0.2.3
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.3
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
oauthlib==3.2.2
olefile==0.46
opencv-python==4.7.0.68
opencv-python-headless==4.9.0.80
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==3.9.1
plotly==5.22.0
portalocker==2.8.2
pprintpp==0.4.0
priority==2.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.21.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==1.0.2
pycairo==1.20.1
pycparser==2.21
pycryptodome==3.14.1
pycryptodomex==3.20.0
pycryptoplus==0.0.1
pycups==2.0.1
pycurl==7.44.1
pyglet==1.5.29
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.7.0
pymacaroons==0.13.0
pymongo==4.7.2
PyNaCl==1.5.0
pynvim==0.4.2
pyOpenSSL==24.1.0
pyparsing==2.4.7
pypdf==3.17.3
pyppmd==1.1.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-ev3dev2==2.1.0.post1
python-json-logger==2.0.4
python-magic==0.4.27
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@f34104cf6ebefacd7b7e07955ee7aaa823e616ac
pytz==2022.1
pyusb==1.2.1
PyWavefront==1.3.3
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.3
pyzstd==0.15.10
qt-range-slider==0.2.7
qtconsole==5.4.0
QtPy==2.3.0
rarfile==4.2
regex==2023.8.8
reportlab==3.6.8
requests==2.26.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.3.3
scikit-image==0.23.2
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.0
shapely==2.0.4
six==1.16.0
smpl==1.3.0
smpl-animation==1.2.1
smpl-debug==1.0.5
smpl-doc==1.1.3
smpl-parallel==1.2.1.1
smpl-util==1.0.10
smpl_io==1.2.0
smplx==0.1.28
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.2
starlette==0.37.2
strawberry-graphql==0.138.1
sympy==1.12
systemd-python==234
tabulate==0.9.0
taskgroup==0.0.0a4
tenacity==8.3.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
texttable==1.6.4
threadpoolctl==3.1.0
tifffile==2024.5.10
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==1.13.1+cpu
torchaudio==0.13.1+cpu
torchvision==0.14.1+cpu
tornado==6.2
tqdm==4.62.3
traitlets==5.8.1
transformers==4.33.2
trimesh==4.4.0
typing_extensions==4.7.1
tzdata==2024.1
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uncertainties==3.1.7
Unidecode==1.3.6
universal-analytics-python3==1.1.1
uri-template==1.2.0
uritemplate==3.0.1
urllib3==1.26.18
vedo==2024.5.1
virtualenv==20.24.2
visualize==0.5.1
voxel51-eta==0.12.6
vtk==9.3.0
wadllib==1.3.6
watchdog==2.1.9
wcwidth==0.2.13
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
Werkzeug==2.1.2
widgetsnbextension==4.0.5
wrapt==1.15.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
xmltodict==0.13.0
yacs==0.1.8
zipp==1.0.0
zstandard==0.15.2
